+++ src/main/java/org/umc/travlocksserver/domain/auth/service/command/OAuthLoginService.java
@@ -10,13 +10,13 @@ import org.umc.travlocksserver.domain.auth.dto.response.AuthOAuthLoginResponseDT
 import org.umc.travlocksserver.domain.auth.entity.OAuthAccount;
 import org.umc.travlocksserver.domain.auth.enums.OAuthProvider;
 import org.umc.travlocksserver.domain.auth.exception.AuthException;
-import org.umc.travlocksserver.domain.auth.exception.code.AuthErrorCode;
+import org.umc.travlocksserver.domain.auth.code.AuthErrorCode;
 import org.umc.travlocksserver.domain.auth.oauth.google.GoogleIdTokenVerifier;
 import org.umc.travlocksserver.domain.auth.repository.OAuthAccountRepository;
 import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.enums.MemberStatus;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
-import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
+import org.umc.travlocksserver.domain.member.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
 import org.umc.travlocksserver.global.profile.DefaultProfileImageProvider;
