+++ src/main/java/org/umc/travlocksserver/domain/member/service/command/MemberOAuthOnboardingService.java
@@ -10,7 +10,7 @@ import org.umc.travlocksserver.domain.member.dto.response.MemberSignupResponseDT
 import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.enums.MemberStatus;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
-import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
+import org.umc.travlocksserver.domain.member.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
 
 @Service
