+++ src/main/java/org/umc/travlocksserver/domain/member/service/command/MemberPasswordUpdateService.java
@@ -6,8 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
-import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
-import org.umc.travlocksserver.domain.member.repository.MemberRepository;
+import org.umc.travlocksserver.domain.member.code.MemberErrorCode;
 
 @Service
 @RequiredArgsConstructor
