+++ src/main/java/org/umc/travlocksserver/domain/template/service/command/TemplateRatingCommandService.java
@@ -3,7 +3,7 @@ package org.umc.travlocksserver.domain.template.service.command;
 import org.springframework.stereotype.Service;
 import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
-import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
+import org.umc.travlocksserver.domain.member.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
 import org.umc.travlocksserver.domain.template.code.TemplateErrorCode;
 import org.umc.travlocksserver.domain.template.dto.request.TemplateRatingCreateRequestDTO;
