+++ src/main/java/org/umc/travlocksserver/global/resolver/LoginUserArgumentResolver.java
@@ -12,7 +12,7 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 import org.umc.travlocksserver.domain.member.entity.Member;
 import org.umc.travlocksserver.domain.member.enums.MemberStatus;
 import org.umc.travlocksserver.domain.member.exception.MemberException;
-import org.umc.travlocksserver.domain.member.exception.code.MemberErrorCode;
+import org.umc.travlocksserver.domain.member.code.MemberErrorCode;
 import org.umc.travlocksserver.domain.member.repository.MemberRepository;
 import org.umc.travlocksserver.global.annotation.LoginUser;
